Move texture utilities to their own file in material loader lib
